共通画面を再作成
