+++ whywhyanalysisApp/Controller/EditAnalysisViewController.swift
@@ -1,0 +1,30 @@
+//
+//  EditAnalysisViewController.swift
+//  whywhyanalysisApp
+//
+//  Created by 長塚かおり on 2020/12/12.
+//  Copyright © 2020 長塚かおり. All rights reserved.
+//
+
+import UIKit
+
+class EditAnalysisViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
